[DT-405] Move categories related code to the categories module
