+++ app/src/main/java/cm/aptoide/pt/analytics/AnalyticsInfoProvider.kt
@@ -1,6 +1,6 @@
 package cm.aptoide.pt.analytics
 
-import cm.aptoide.pt.feature_apps.data.network.analytics.AptoideAnalyticsInfoProvider
+import cm.aptoide.pt.feature_categories.analytics.AptoideAnalyticsInfoProvider
 
 class AnalyticsInfoProvider : AptoideAnalyticsInfoProvider {
   override suspend fun getAnalyticsId(): String? = null
