+++ app/src/main/java/cm/aptoide/pt/di/AppModule.kt
@@ -1,8 +1,8 @@
 package cm.aptoide.pt.di
 
 import cm.aptoide.pt.analytics.AnalyticsInfoProvider
-import cm.aptoide.pt.feature_apps.data.network.analytics.AptoideAnalyticsInfoProvider
-import cm.aptoide.pt.feature_apps.data.network.firebase.AptoideFirebaseInfoProvider
+import cm.aptoide.pt.feature_categories.analytics.AptoideAnalyticsInfoProvider
+import cm.aptoide.pt.feature_categories.analytics.AptoideFirebaseInfoProvider
 import cm.aptoide.pt.firebase.FirebaseInfoProvider
 import dagger.Module
 import dagger.Provides
