+++ app/src/main/java/cm/aptoide/pt/firebase/FirebaseInfoProvider.kt
@@ -1,6 +1,6 @@
 package cm.aptoide.pt.firebase
 
-import cm.aptoide.pt.feature_apps.data.network.firebase.AptoideFirebaseInfoProvider
+import cm.aptoide.pt.feature_categories.analytics.AptoideFirebaseInfoProvider
 
 class FirebaseInfoProvider : AptoideFirebaseInfoProvider {
   override suspend fun getFirebaseToken(): String? = null
