--- feature_apps/src/main/java/cm/aptoide/pt/feature_apps/data/AppCategory.kt
@@ -1,3 +1,0 @@
-package cm.aptoide.pt.feature_apps.data
-
-data class AppCategory(val name: String, val type: String)
