+++ feature_apps/src/main/java/cm/aptoide/pt/feature_apps/data/AppsRepository.kt
@@ -16,5 +16,4 @@ interface AppsRepository {
 
   fun getAppVersions(packageName: String): Flow<List<App>>
 
-  suspend fun getAppsCategories(packageNames: List<String>): List<AppCategory>
 }
