+++ feature_apps/src/main/java/cm/aptoide/pt/feature_apps/data/di/RepositoryModule.kt
@@ -3,8 +3,6 @@ package cm.aptoide.pt.feature_apps.data.di
 import cm.aptoide.pt.aptoide_network.di.RetrofitV7
 import cm.aptoide.pt.aptoide_network.di.StoreName
 import cm.aptoide.pt.feature_apps.data.*
-import cm.aptoide.pt.feature_apps.data.network.analytics.AptoideAnalyticsInfoProvider
-import cm.aptoide.pt.feature_apps.data.network.firebase.AptoideFirebaseInfoProvider
 import cm.aptoide.pt.feature_apps.data.network.service.AppsRemoteService
 import cm.aptoide.pt.feature_apps.data.network.service.AptoideAppsNetworkService
 import cm.aptoide.pt.feature_campaigns.CampaignRepository
@@ -35,14 +33,10 @@ internal object RepositoryModule {
   fun providesAppsRemoteService(
     @RetrofitV7 retrofitV7: Retrofit,
     @StoreName storeName: String,
-    analyticsInfoProvider: AptoideAnalyticsInfoProvider,
-    messagingInfoProvider: AptoideFirebaseInfoProvider
   ): AppsRemoteService {
     return AptoideAppsNetworkService(
       appsRemoteDataSource = retrofitV7.create(AptoideAppsNetworkService.Retrofit::class.java),
-      storeName = storeName,
-      analyticsInfoProvider = analyticsInfoProvider,
-      messagingInfoProvider = messagingInfoProvider
+      storeName = storeName
     )
   }
 }
