--- feature_apps/src/main/java/cm/aptoide/pt/feature_apps/data/network/model/AppCategoryJSON.kt
@@ -1,6 +1,0 @@
-package cm.aptoide.pt.feature_apps.data.network.model
-
-import androidx.annotation.Keep
-
-@Keep
-data class AppCategoryJSON(val name: String, val type: String)
