+++ feature_apps/src/main/java/cm/aptoide/pt/feature_apps/data/network/service/AppsRemoteService.kt
@@ -2,7 +2,6 @@ package cm.aptoide.pt.feature_apps.data.network.service
 
 import cm.aptoide.pt.aptoide_network.data.network.base_response.BaseV7DataListResponse
 import cm.aptoide.pt.aptoide_network.data.network.base_response.BaseV7ListResponse
-import cm.aptoide.pt.feature_apps.data.network.model.AppCategoryJSON
 import cm.aptoide.pt.feature_apps.data.network.model.AppJSON
 import cm.aptoide.pt.feature_apps.data.network.model.GetAppResponse
 
@@ -17,6 +16,4 @@ internal interface AppsRemoteService {
   suspend fun getRecommended(url: String, bypassCache: Boolean = false): BaseV7DataListResponse<AppJSON>
 
   suspend fun getAppVersionsList(packageName: String): BaseV7ListResponse<AppJSON>
-
-  suspend fun getAppCategories(packageNames: List<String>): BaseV7ListResponse<AppCategoryJSON>
 }
