+++ feature_apps/src/main/java/cm/aptoide/pt/feature_apps/data/network/service/AptoideAppsNetworkService.kt
@@ -4,16 +4,12 @@ import cm.aptoide.pt.aptoide_network.data.network.CacheConstants.CACHE_CONTROL_H
 import cm.aptoide.pt.aptoide_network.data.network.CacheConstants.NO_CACHE
 import cm.aptoide.pt.aptoide_network.data.network.base_response.BaseV7DataListResponse
 import cm.aptoide.pt.aptoide_network.data.network.base_response.BaseV7ListResponse
-import cm.aptoide.pt.feature_apps.data.network.analytics.AptoideAnalyticsInfoProvider
-import cm.aptoide.pt.feature_apps.data.network.firebase.AptoideFirebaseInfoProvider
 import cm.aptoide.pt.feature_apps.data.network.model.*
 import retrofit2.http.*
 
 internal class AptoideAppsNetworkService(
   private val appsRemoteDataSource: Retrofit,
-  private val storeName: String,
-  private val analyticsInfoProvider: AptoideAnalyticsInfoProvider,
-  private val messagingInfoProvider: AptoideFirebaseInfoProvider,
+  private val storeName: String
 ) :
   AppsRemoteService {
   override suspend fun getAppsList(
@@ -61,17 +57,6 @@ internal class AptoideAppsNetworkService(
     return appsRemoteDataSource.getAppVersionsList(packageName, storeName)
   }
 
-  override suspend fun getAppCategories(packageNames: List<String>): BaseV7ListResponse<AppCategoryJSON> {
-    val analyticsId = analyticsInfoProvider.getAnalyticsId()
-    val firebaseToken = messagingInfoProvider.getFirebaseToken()
-    return appsRemoteDataSource.getAppsCategories(
-      names = Names(packageNames),
-      storeName = storeName,
-      analyticsId = analyticsId,
-      firebaseToken = firebaseToken
-    )
-  }
-
   internal interface Retrofit {
     @GET("apps/get/{query}")
     suspend fun getAppsList(
@@ -111,13 +96,5 @@ internal class AptoideAppsNetworkService(
       @Query("store_name") storeName: String,
       @Query("aab") aab: Int = 1,
     ): BaseV7ListResponse<AppJSON>
-
-    @POST("hub/apps/get/")
-    suspend fun getAppsCategories(
-      @Query("user_uid") analyticsId: String?,
-      @Query("store_name") storeName: String,
-      @Query("firebase_token") firebaseToken: String?,
-      @Body names: Names
-    ): BaseV7ListResponse<AppCategoryJSON>
   }
 }
