+++ feature_categories/src/main/java/cm/aptoide/pt/feature_categories/data/CategoriesRepository.kt
@@ -1,5 +1,6 @@
 package cm.aptoide.pt.feature_categories.data
 
+import cm.aptoide.pt.feature_categories.domain.AppCategory
 import cm.aptoide.pt.feature_categories.domain.Category
 import kotlinx.coroutines.flow.Flow
 
@@ -8,4 +9,6 @@ interface CategoriesRepository {
   fun getCategoriesList(url: String): Flow<List<Category>>
 
   fun getHomeBundleActionListCategories(bundleTag: String): Flow<Pair<List<Category>, String>>
+
+  suspend fun getAppsCategories(packageNames: List<String>): List<AppCategory>
 }
