+++ feature_categories/src/main/java/cm/aptoide/pt/feature_categories/data/model/AppCategoryJSON.kt
@@ -1,0 +1,6 @@
+package cm.aptoide.pt.feature_categories.data.model
+
+import androidx.annotation.Keep
+
+@Keep
+data class AppCategoryJSON(val name: String, val type: String)
