+++ feature_categories/src/main/java/cm/aptoide/pt/feature_categories/di/RepositoryModule.kt
@@ -1,7 +1,9 @@
-package cm.aptoide.pt.feature_categories.data.di
+package cm.aptoide.pt.feature_categories.di
 
 import cm.aptoide.pt.aptoide_network.di.RetrofitV7
 import cm.aptoide.pt.aptoide_network.di.StoreName
+import cm.aptoide.pt.feature_categories.analytics.AptoideAnalyticsInfoProvider
+import cm.aptoide.pt.feature_categories.analytics.AptoideFirebaseInfoProvider
 import cm.aptoide.pt.feature_categories.data.AptoideCategoriesRepository
 import cm.aptoide.pt.feature_categories.data.CategoriesRepository
 import cm.aptoide.pt.feature_home.data.WidgetsRepository
@@ -21,12 +23,16 @@ internal object RepositoryModule {
   fun providesCategoriesRepository(
     @RetrofitV7 retrofitV7: Retrofit,
     @StoreName storeName: String,
-    widgetsRepository: WidgetsRepository
+    widgetsRepository: WidgetsRepository,
+    analyticsInfoProvider: AptoideAnalyticsInfoProvider,
+    messagingInfoProvider: AptoideFirebaseInfoProvider
   ): CategoriesRepository {
     return AptoideCategoriesRepository(
       widgetsRepository = widgetsRepository,
       categoriesRemoteDataSource = retrofitV7.create(AptoideCategoriesRepository.Retrofit::class.java),
-      storeName = storeName
+      storeName = storeName,
+      analyticsInfoProvider = analyticsInfoProvider,
+      messagingInfoProvider = messagingInfoProvider
     )
   }
 }
