+++ feature_categories/src/main/java/cm/aptoide/pt/feature_categories/domain/AppCategory.kt
@@ -1,0 +1,3 @@
+package cm.aptoide.pt.feature_categories.domain
+
+data class AppCategory(val name: String, val type: String)
